+++ cmd/defaults.go
@@ -100,13 +100,7 @@ func updateDefaults(cmd *cobra.Command, args []string) {
 	defaults.ShowDefaults(defaults.Defaults())
 }
 
-func enableBashCompletion(cmd *cobra.Command, args []string) {
-
-	flags := cmd.Flags()
-	useRemote, _ := flags.GetBool(globals.RemoteLabel)
-	runIt, _ := flags.GetBool(globals.RunItLabel)
-	remoteUrl, _ := flags.GetString(globals.RemoteUrlLabel)
-	completionFile, _ := flags.GetString(globals.CompletionFileLabel)
+func processBashCompletionEnabling(useRemote, runIt bool, remoteUrl, completionFile string) error {
 	useLocal := completionFile != ""
 
 	bashCompletionScript := path.Join("/etc", "bash_completion")
@@ -117,21 +111,21 @@ func enableBashCompletion(cmd *cobra.Command, args []string) {
 		if common.DirExists(alternateDestinationDir) {
 			destinationDir = alternateDestinationDir
 		} else {
-			common.Exitf(1, "neither %s or %s found", destinationDir, alternateDestinationDir)
+			return fmt.Errorf("neither %s or %s found", destinationDir, alternateDestinationDir)
 		}
 	}
 	if !common.FileExists(bashCompletionScript) {
 		if common.FileExists(bashCompletionScript) {
 			bashCompletionScript = alternateBashCompletionScript
 		} else {
-			common.Exitf(1, "neither %s or %s found", bashCompletionScript, alternateBashCompletionScript)
+			return fmt.Errorf("neither %s or %s found", bashCompletionScript, alternateBashCompletionScript)
 		}
 	}
 	if completionFile == "" {
 		completionFile = globals.CompletionFileValue
 	}
 	if useLocal && useRemote {
-		common.Exitf(1, "Only one of '--%s' or '--%s' should be used", globals.CompletionFileValue, globals.RemoteLabel)
+		return fmt.Errorf("only one of '--%s' or '--%s' should be used", globals.CompletionFileValue, globals.RemoteLabel)
 	}
 	if !useRemote {
 		useLocal = true
@@ -150,20 +144,20 @@ func enableBashCompletion(cmd *cobra.Command, args []string) {
 	}
 	if useRemote {
 		if remoteUrl == "" {
-			common.Exitf(1, "Remote URL at '--%s' cannot be empty", globals.RemoteUrlLabel)
+			return fmt.Errorf("remote URL at '--%s' cannot be empty", globals.RemoteUrlLabel)
 		}
 		if common.FileExists(completionFile) {
-			common.Exitf(1, globals.ErrFileAlreadyExists, completionFile)
+			return fmt.Errorf(globals.ErrFileAlreadyExists, completionFile)
 		}
 		err := rest.DownloadFile(completionFile, remoteUrl, true, globals.MB)
 		if err != nil {
-			common.Exitf(1, "error downloading %s: %s", completionFile, err)
+			return fmt.Errorf("error downloading %s: %s", completionFile, err)
 		}
 		fmt.Printf("Download of file %s was successful\n", completionFile)
 	}
 
 	if !common.FileExists(completionFile) {
-		common.Exitf(1, globals.ErrFileNotFound, completionFile)
+		return fmt.Errorf(globals.ErrFileNotFound, completionFile)
 	}
 
 	fmt.Printf("# completion file: %s\n", completionFile)
@@ -173,15 +167,15 @@ func enableBashCompletion(cmd *cobra.Command, args []string) {
 		// Get the checksum of both files, so we can skip the copy if they are already the same
 		sourceChecksum, err := common.GetFileSha256(completionFile)
 		if err != nil {
-			common.Exitf(1, "error getting checksum from file %s", completionFile)
+			return fmt.Errorf("error getting checksum from file %s", completionFile)
 		}
 		destChecksum, err := common.GetFileSha256(destinationFile)
 		if err != nil {
-			common.Exitf(1, "error getting checksum from file %s", destinationFile)
+			return fmt.Errorf("error getting checksum from file %s", destinationFile)
 		}
 		if sourceChecksum == destChecksum {
 			fmt.Printf("Files '%s' and '%s' have the same checksum - Copy is not needed\n", completionFile, destinationFile)
-			return
+			return nil
 		}
 	}
 
@@ -191,10 +185,10 @@ func enableBashCompletion(cmd *cobra.Command, args []string) {
 		output, err := common.RunCmdWithArgs("sudo", []string{"cp", completionFile, destinationDir})
 		if err != nil {
 			fmt.Printf("%s\n", output)
-			common.Exitf(1, "error copying bash completion file into %s: %s", destinationDir, err)
+			return fmt.Errorf("error copying bash completion file into %s: %s", destinationDir, err)
 		}
 		if !common.FileExists(destinationFile) {
-			common.Exitf(1, "error after copying bash completion file: "+globals.ErrFileNotFound, destinationFile)
+			return fmt.Errorf("error after copying bash completion file: "+globals.ErrFileNotFound, destinationFile)
 		}
 		fmt.Printf("# File copied to %s\n", destinationFile)
 	} else {
@@ -202,6 +196,20 @@ func enableBashCompletion(cmd *cobra.Command, args []string) {
 		fmt.Printf("# Run the command: sudo cp %s %s\n", completionFile, destinationDir)
 	}
 	fmt.Printf("# Run the command 'source %s'\n", bashCompletionScript)
+	return nil
+}
+
+func enableBashCompletion(cmd *cobra.Command, args []string) {
+
+	flags := cmd.Flags()
+	useRemote, _ := flags.GetBool(globals.RemoteLabel)
+	runIt, _ := flags.GetBool(globals.RunItLabel)
+	remoteUrl, _ := flags.GetString(globals.RemoteUrlLabel)
+	completionFile, _ := flags.GetString(globals.CompletionFileLabel)
+	err := processBashCompletionEnabling(useRemote, runIt, remoteUrl, completionFile)
+	if err != nil {
+		common.Exitf(1, "%s", err)
+	}
 }
 
 func showFlagAliases(cmd *cobra.Command, args []string) {